requirements.txt
RPi.GPIO==0.7.0
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
gunicorn==20.0.4
idna==2.10
multidict==5.0.2
typing-extensions==3.7.4.3
uvloop==0.14.0
yarl==1.6.3
